Will move 'LocalFeedItem' to its own file
